Adding event on submit survey

+++ app/Http/Controllers/LSurveyController.php
@@ -47,17 +47,16 @@ class LSurveyController extends Controller
             $LastNameAPI  = $contact->lastname;
             $FirstNameAPI = $contact->firstname;
 
-            $limeConnection = self::connectLime(); //Start limesurvey session
+            $limeConnection  = self::connectLime(); //Start limesurvey session
             $myJSONRPCClient = $limeConnection['myJSONRPCClient'];
-            $sSessionKey = $limeConnection['sessionKey'];
-
+            $sSessionKey     = $limeConnection['sessionKey'];
             // Define the token params
             $tokenParams = array(
-                "email" => $email,
-                "lastname" => $LastNameAPI,
-                "firstname" => $FirstNameAPI,
-                "token" => $token,
-                // "language" => 'en',
+                "email"       => $email,
+                "lastname"    => $LastNameAPI,
+                "firstname"   => $FirstNameAPI,
+                "token"       => $token,
+                // "language"    => 'es-MX',
                 "emailstatus" => "OK"
             );
 
@@ -65,7 +64,6 @@ class LSurveyController extends Controller
             $bCreateToken = true;
             // Create the tokens
             $newToken = $myJSONRPCClient->add_participants($sSessionKey, $iSurveyID, $aParticipantData, $bCreateToken);
-
             // return $newToken[0]['token'];
             if (array_key_exists("status", $newToken)) { // If catched survey error starus
                 return response()->json(["error" => [$newToken, $sSessionKey]]);
@@ -74,14 +72,13 @@ class LSurveyController extends Controller
                 $newMail = $myJSONRPCClient->invite_participants($sSessionKey, $iSurveyID, $tokenIDs, true);
                 // Print returned results
                 $surveyValues = [];
-
-
                 if ($newMail[$newToken[0]['tid']]['status'] == 'OK') {
                     array_push(
                         $surveyValues,
                         $iSurveyID,
                         $newToken[0]['token'],
-                        $newToken[0]['tid']
+                        $newToken[0]['tid'],
+                        $newToken[0]['language']
                     );
                 }
 
@@ -92,7 +89,7 @@ class LSurveyController extends Controller
                     $clitem =  $vtiger->search($clitemQuery)->result[0];
                     $obj = $vtiger->retrieve($clitem->id);
                     //$this->consoleWrite()->writeln($limeConnection['lime_connection']['LS_BASEURL'] . '/' . $surveyValues[0] . "?token=" . $surveyValues[1] . "&lang=" . "en");
-                    $obj->result->cf_1212 = $limeConnection['lime_connection']['LS_BASEURL'] . '/' . $surveyValues[0] . "?token=" . $surveyValues[1] . "&lang=" . "en"; //help link
+                    $obj->result->cf_1212 = $limeConnection['lime_connection']['LS_BASEURL'] . '/' . $surveyValues[0] . "?token=" . $surveyValues[1] . "&lang=" .  "en"; //help link
                     $obj->result->cf_acf_rtf_1208 = "This clitem is a questionaire";
                     $vtiger->update($obj->result);
                     //return [$clitem, $obj];
@@ -119,28 +116,26 @@ class LSurveyController extends Controller
     public function exportResponse(Request $request)
     {
         try {
-            $user      = Auth::user();
             $vtiger    = new Vtiger();
             $task      = new CloneDBController;
             $docsTask  = new DocumentController();
             $now       = Carbon::now()->format('H:i:s');
             $received  = false;
 
+            $this->consoleWrite()->writeln("Here we go 1");
+
             if ($request->form != 'vue') {
                 $request->form = '';
             }
-            $oncpItem = CLItem::where("clitemsno", $request->clitemsno)->firstOrFail();
-
-            if (!$oncpItem) {
-                return 404;
+            if ($request->survey_url) {
+                $oncpItem = CLItem::where("cf_1212", "like", "$request->survey_url" . "%")->firstOrFail();
+                $clitemQuery = DB::table('CLItems')->select('*')->where("cf_1212", "like", "$request->survey_url" . "%")->take(1);
+            } else {
+                $oncpItem = CLItem::where("clitemsno", $request->clitemsno)->firstOrFail();
+                $clitemQuery = DB::table('CLItems')->select('*')->where("clitemsno", $request->clitemsno)->take(1);
             }
             /* Find clitem of current survey */
-            $clitemQuery = DB::table('CLItems')->select('*')
-                ->where("clitemsno", $request->clitemsno)->take(1);
-
             $clitem = $vtiger->search($clitemQuery);
-
-            // return  empty($clitem->result) ? 'true' : 'false';
 
             if ($clitem->success === false || empty($clitem->result)) {
                 return response()->json("Item not found", 404);
@@ -162,6 +157,8 @@ class LSurveyController extends Controller
             $urlObj = parse_url($clitem->cf_1212);
             $iSurveyID = str_replace('/', '', $urlObj['path']);
 
+            $this->consoleWrite()->writeln($clitem->cf_1212);
+
             $urlQuery = $urlObj['query'];
             $urlQueryAsArr = [];
 
@@ -169,6 +166,7 @@ class LSurveyController extends Controller
             $sToken = $urlQueryAsArr['token'];
 
             $limeConnection = self::connectLime(); //Start limesurvey session
+
 
             $myJSONRPCClient = $limeConnection['myJSONRPCClient'];
             $sSessionKey = $limeConnection['sessionKey'];
@@ -210,7 +208,11 @@ class LSurveyController extends Controller
                 $file = str_replace(' ', '_', $clitem->name);
                 $file = $file . '.pdf';
 
+                $this->consoleWrite()->writeln("Here we go 2");
+                $this->consoleWrite()->writeln($surveyResponses[0]->submitdate);
+
                 if ($surveyResponses[0]->submitdate != null) {
+                    $this->consoleWrite()->writeln("Here we go 3");
                     if (!Storage::exists($directory . '/' . $file)) {
                         if (env('APP_ENV') === 'local') {
                             self::download('', '../storage/app/' . $directory . '/' . $file, $exportSurveyAsPDF); //save survey file on storage folder
@@ -219,38 +221,51 @@ class LSurveyController extends Controller
                         }
                     }
                     if (Storage::exists($directory . '/' . $file)) {
+
+                        if (env('APP_ENV') === 'local') {
+                            $this->consoleWrite()->writeln("just here");
+                        }
+
+
                         $obj = $vtiger->retrieve($clitem->id);
-                        $request->request->add(['cid'       => $user->vtiger_contact_id]);
+
+                        $this->consoleWrite()->writeln($obj->result->clitemsno);
+
+                        $request->request->add(['cid'       => $contact->contact_no]);
                         $request->request->add(['case'      => $case->ticket_no . '-' . $case->ticketcategories]);
                         $request->request->add(['checklist' => $checklist->checklistno . '-' . $checklist->cf_1706]);
                         $request->request->add(['clitem'    => $clitem->clitemsno . '-' . $clitem->cf_1200]);
 
                         $files =  $docsTask->checkDocuments($request);
-                        $arrAsStr = implode(', ', $files);
+                        $arrAsStr =  $files ? implode(', ', $files) : '';
+
 
                         $obj->result->description = "File uploaded at: " . $now;
                         $obj->result->cf_2370   = $arrAsStr; //set on metadata field
                         $obj->result->cf_1214     = "$contact->cf_1332/$contact->contact_no/$contact->contact_no-cases/$case->ticket_no-$case->ticketcategories/01_SuppliedDocs"; //GD Link
-                        $vtiger->update($obj->result);
+                        $this->consoleWrite()->writeln("try to update");
 
+                        $vtiger->update($obj->result);
+                        
+                        $this->consoleWrite()->writeln("setted");
+                        return "any";
                         sleep(8); // wait 12 seconds
 
                         $obj2  = $vtiger->retrieve($clitem->id);
-                        /*   if (env('APP_ENV') === 'local') {
-                            $this->consoleWrite()->writeln('item updated');
-                            $this->consoleWrite()->writeln($updatedItem->cf_1578);
-                        } */
-
                         //call destroller
                         do {
                             if ($obj2->result->cf_1578 === "Received") {
+                                $request->request->add(['checklist_id' => $checklist->checklistno]);
+
                                 $updatedItem = $task->updateCLItemFromImmcase($request);
                                 $task->updateChecklistFromImmcase($request);
 
-                                $urls = explode(', ', $updatedItem->cf_2370);
-                                foreach ($urls as  $url) {
-                                    $request->request->add(['file' => $url]);
-                                    $docsTask->destroy($request);
+                                if ($updatedItem->cf_2370 != '') {
+                                    $urls = explode(', ', $updatedItem->cf_2370);
+                                    foreach ($urls as  $url) {
+                                        $request->request->add(['file' => $url]);
+                                        $docsTask->destroy($request);
+                                    }
                                 }
                             } else {
                                 sleep(3);
@@ -259,6 +274,10 @@ class LSurveyController extends Controller
                         } while ($obj2->result->cf_1578 != "Received");
                         $received = true;
                     }
+                }
+            } else {
+                if (env('APP_ENV') === 'local') {
+                    $this->consoleWrite()->writeln("Cant export");
                 }
             }
             // Release the session key
@@ -293,6 +312,9 @@ class LSurveyController extends Controller
         $clitem = $vtiger->search($clitemQuery);
 
         if ($clitem->success === false) {
+            if (env('APP_ENV') === 'local') {
+                $this->consoleWrite()->writeln('error at 302');
+            }
             return response()->json("Item not found", 404);
         }
         $clitem =  $clitem->result[0];
@@ -303,7 +325,7 @@ class LSurveyController extends Controller
             $clitem = $vtiger->search($clitemQuery);
         }
 
-        //limesurvey connection 
+        //limesurvey connection
         $urlObj = parse_url($clitem->cf_1212);
         $iSurveyID = str_replace('/', '', $urlObj['path']);
 
@@ -329,7 +351,7 @@ class LSurveyController extends Controller
             $aFields = null
         );
 
-        $surveyResponses =[];
+        $surveyResponses = [];
         if (is_string($exportSurvey)) {
             $surveyResults = json_decode(base64_decode($exportSurvey));
             $surveyResponses =  $surveyResults->responses;
@@ -398,5 +420,29 @@ class LSurveyController extends Controller
         /*  if (file_exists($filepath)) {
             unlink($filepath);
         } */
+    }
+
+    /**
+     * It method be automatic called at sbmit a survey by an participant
+     */
+    public function onSubmit(Request $request, $id, $tkn, $ln)
+    {
+        $link = "http://forms.gercanada.com/$id?token=$tkn";
+        // $link = "https://forms.gercanada.com/$id?token=$tkn&lang=$ln";
+        try {
+            $request->request->add(['survey_url' => $link]);
+            $task = new LSurveyController();
+            $exported = $task->exportResponse($request);
+
+            return view('checklists.items.survey.done');
+            // return $exported;
+        } catch (Exception $e) {
+            return $this->returnJsonError($e, ['LSurveyController' => 'onSubmit']);
+        }
+
+        /*   http: //forms.gercanada.com/1000?token=H1AC0X6X6eanyYy&lang=en
+        http: //forms.gercanada.com/{SAVEDID}?token={TOKEN}&lang={LANG}
+        https: //70b6-187-212-180-149.ngrok.io/submitsurvey/helplink/http://forms.gercanada.com/{SAVEDID}?token={TOKEN}&lang={LANG}
+         https: //70b6-187-212-180-149.ngrok.io/submitsurvey/id{SAVEDID}/tkn/{TOKEN}/ln/{LANG} */
     }
 }
